+++ src/main/java/wooteco/subway/DataLoader.java
@@ -3,13 +3,13 @@ package wooteco.subway;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
-import wooteco.subway.line.dao.LineDao;
-import wooteco.subway.line.dao.SectionDao;
+import wooteco.subway.line.infrastructure.dao.LineDao;
+import wooteco.subway.line.infrastructure.dao.SectionDao;
 import wooteco.subway.line.domain.Line;
 import wooteco.subway.line.domain.Section;
-import wooteco.subway.member.dao.MemberDao;
+import wooteco.subway.member.infrastructure.dao.MemberDao;
 import wooteco.subway.member.domain.Member;
-import wooteco.subway.station.dao.StationDao;
+import wooteco.subway.station.infrastructure.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 
 @Component
