+++ src/main/java/wooteco/subway/auth/application/AuthService.java
@@ -2,10 +2,10 @@ package wooteco.subway.auth.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.subway.auth.dto.TokenRequest;
-import wooteco.subway.auth.dto.TokenResponse;
+import wooteco.subway.auth.ui.dto.TokenRequest;
+import wooteco.subway.auth.ui.dto.TokenResponse;
 import wooteco.subway.auth.infrastructure.JwtTokenProvider;
-import wooteco.subway.member.dao.MemberDao;
+import wooteco.subway.member.infrastructure.dao.MemberDao;
 import wooteco.subway.member.domain.LoginMember;
 import wooteco.subway.member.domain.Member;
