+++ src/main/java/wooteco/subway/auth/ui/AuthController.java
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.auth.application.AuthService;
-import wooteco.subway.auth.dto.TokenRequest;
-import wooteco.subway.auth.dto.TokenResponse;
+import wooteco.subway.auth.ui.dto.TokenRequest;
+import wooteco.subway.auth.ui.dto.TokenResponse;
 
 @RestController
 @CrossOrigin(origins = "*", allowedHeaders = "*")
