+++ src/main/java/wooteco/subway/auth/ui/dto/TokenRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.auth.dto;
+package wooteco.subway.auth.ui.dto;
 
 public class TokenRequest {
     private String email;
