+++ src/main/java/wooteco/subway/line/ui/dto/LineRequest.java
@@ -1,17 +1,29 @@
-package wooteco.subway.line.dto;
+package wooteco.subway.line.ui.dto;
 
-public class SectionRequest {
+public class LineRequest {
+    private String name;
+    private String color;
     private Long upStationId;
     private Long downStationId;
     private int distance;
 
-    public SectionRequest() {
+    public LineRequest() {
     }
 
-    public SectionRequest(Long upStationId, Long downStationId, int distance) {
+    public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
+        this.name = name;
+        this.color = color;
         this.upStationId = upStationId;
         this.downStationId = downStationId;
         this.distance = distance;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
     }
 
     public Long getUpStationId() {
