+++ src/main/java/wooteco/subway/line/ui/dto/LineResponse.java
@@ -1,7 +1,7 @@
-package wooteco.subway.line.dto;
+package wooteco.subway.line.ui.dto;
 
 import wooteco.subway.line.domain.Line;
-import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.ui.dto.StationResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
