+++ src/main/java/wooteco/subway/member/ui/dto/MemberRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.member.dto;
+package wooteco.subway.member.ui.dto;
 
 import wooteco.subway.member.domain.Member;
