+++ src/main/java/wooteco/subway/path/application/PathService.java
@@ -4,10 +4,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.line.application.LineService;
 import wooteco.subway.line.domain.Line;
-import wooteco.subway.member.domain.LoginMember;
 import wooteco.subway.path.domain.SubwayPath;
-import wooteco.subway.path.dto.PathResponse;
-import wooteco.subway.path.dto.PathResponseAssembler;
+import wooteco.subway.path.ui.dto.PathResponse;
+import wooteco.subway.path.ui.dto.PathResponseAssembler;
 import wooteco.subway.station.application.StationService;
 import wooteco.subway.station.domain.Station;
