+++ src/main/java/wooteco/subway/path/ui/PathController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.path.application.PathService;
-import wooteco.subway.path.dto.PathResponse;
+import wooteco.subway.path.ui.dto.PathResponse;
 
 @RestController
 @CrossOrigin(origins = "*", allowedHeaders = "*")
