+++ src/main/java/wooteco/subway/path/ui/dto/PathResponseAssembler.java
@@ -1,7 +1,7 @@
-package wooteco.subway.path.dto;
+package wooteco.subway.path.ui.dto;
 
 import wooteco.subway.path.domain.SubwayPath;
-import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.ui.dto.StationResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
