+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,10 +1,10 @@
 package wooteco.subway.station.application;
 
 import org.springframework.stereotype.Service;
-import wooteco.subway.station.dao.StationDao;
+import wooteco.subway.station.infrastructure.dao.StationDao;
 import wooteco.subway.station.domain.Station;
-import wooteco.subway.station.dto.StationRequest;
-import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.ui.dto.StationRequest;
+import wooteco.subway.station.ui.dto.StationResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
