+++ src/main/java/wooteco/subway/station/infrastructure/dao/StationDao.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station.dao;
+package wooteco.subway.station.infrastructure.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
