+++ src/main/java/wooteco/subway/station/ui/StationController.java
@@ -4,8 +4,8 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import wooteco.subway.station.application.StationService;
-import wooteco.subway.station.dto.StationRequest;
-import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.ui.dto.StationRequest;
+import wooteco.subway.station.ui.dto.StationResponse;
 
 import java.net.URI;
 import java.sql.SQLException;
