+++ src/main/java/wooteco/subway/station/ui/dto/StationResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station.dto;
+package wooteco.subway.station.ui.dto;
 
 import wooteco.subway.station.domain.Station;
