+++ src/test/java/wooteco/subway/station/StationAcceptanceTest.java
@@ -8,8 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.AcceptanceTest;
-import wooteco.subway.station.dto.StationRequest;
-import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.ui.dto.StationRequest;
+import wooteco.subway.station.ui.dto.StationResponse;
 
 import java.util.Arrays;
 import java.util.List;
